mouse events have x,y relative to window's position

+++ userspace/WindowServer/Window.h
@@ -13,8 +13,8 @@ public:
     u32 buff_guid() const {return m_buff_guid;}
     void* buff_addr() const {return m_buff_addr;}
     u32 buff_size() const {return m_buff_size;}
-    u32 x() const {return m_x;}
-    u32 y() const {return m_y;}
+    u16 x() const {return m_x;}
+    u16 y() const {return m_y;}
     u32 width() const {return m_width;}
     u32 height() const {return m_height;}
     Rectangle rectangle() const;
